==> assessor_land_exploration/California/Los Angeles County/notebook/01_transformation.ipynb <==
## Data Exploration of Hazards, Local Roll, Secured File, and Transactions of Parcels within Southern California

#### LA County

%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np
# Set PD display options to show all columns and column width
pd.set_option('display.max_columns', None)
pd.set_option('display.max_colwidth', 1000)
pd.set_option('display.max_rows', 500)
pd.set_option('display.width', 1000)


### Load in the data from Dropbox

DS_1 = pd.read_csv('../data/dropbox/DS04 Part 1 Data Sales.csv')
DS_2 = pd.read_csv('../data/dropbox/DS04 Part 2 Data Sales.csv')
DS_3 = pd.read_csv('../data/dropbox/DS04 Part 3 Data Sales.csv')

local_roll = pd.read_excel('../data/dropbox/Local Roll Part 1 Data Sales.xlsx')
local_roll_2 = pd.read_excel('../data/dropbox/Local Roll Part 2 Data Sales.xlsx')
local_roll_3 = pd.read_excel('../data/dropbox/Local Roll Part 3 Data Sales.xlsx')
local_roll_4 = pd.read_excel('../data/dropbox/Local Roll Part 4 Data Sales.xlsx')
local_roll_5 = pd.read_excel('../data/dropbox/Local Roll Part 5 Data Sales.xlsx')

sales_list = pd.read_excel('../data/dropbox/SALESLIST_012424 Data Sales.xlsx')

# Find all First Owner Name where "SHIOTA" is in the name
#DS_3[DS_3['First Owner Name'].str.contains('SHIOTA', na=False)]

DS_1.rename(columns=lambda x: x.strip(), inplace=True)
DS_2.rename(columns=lambda x: x.strip(), inplace=True)
DS_3.rename(columns=lambda x: x.strip(), inplace=True)

local_roll.rename(columns=lambda x: x.strip(), inplace=True)
local_roll_2.rename(columns=lambda x: x.strip(), inplace=True)
local_roll_3.rename(columns=lambda x: x.strip(), inplace=True)
local_roll_4.rename(columns=lambda x: x.strip(), inplace=True)
local_roll_5.rename(columns=lambda x: x.strip(), inplace=True)

sales_list.rename(columns=lambda x: x.strip(), inplace=True)

DS_1 = DS_1.map(lambda x: x.strip() if isinstance(x, str) else x)
DS_2 = DS_2.map(lambda x: x.strip() if isinstance(x, str) else x)
DS_3 = DS_3.map(lambda x: x.strip() if isinstance(x, str) else x)

local_roll = local_roll.map(lambda x: x.strip() if isinstance(x, str) else x)
local_roll_2 = local_roll_2.map(lambda x: x.strip() if isinstance(x, str) else x)
local_roll_3 = local_roll_3.map(lambda x: x.strip() if isinstance(x, str) else x)
local_roll_4 = local_roll_4.map(lambda x: x.strip() if isinstance(x, str) else x)
local_roll_5 = local_roll_5.map(lambda x: x.strip() if isinstance(x, str) else x)

sales_list.rename(columns=lambda x: x.strip(), inplace=True)

# Concatenate the three datasets
DS = pd.concat([DS_1, DS_2, DS_3], axis=0)

# Concatenate the local roll datasets
local_roll = pd.concat([local_roll, local_roll_2, local_roll_3, local_roll_4, local_roll_5], axis=0)

# Save to parquet in transformed directory
DS.to_csv('../data/transformed/DS_transformed.csv', index=False)
local_roll.to_csv('../data/transformed/local_roll_transformed.csv', index=False)
sales_list.to_csv('../data/transformed/sales_list_transformed.csv', index=False)
