==> assessor_land_exploration/California/Los Angeles County/notebook/05_map_data_dictionaries.ipynb <==
# Notebook 5: This notebook identifies the land use of a given location. 
### Load in the data that contains the columns "land_use". 
### Map the codes according to the use_code_2023.py file containing the dictionary of unique land codes

# Grab the data from the FINAL DATA DIR
import os
import sys
# Get the current working directory
current_directory = os.getcwd()
parent_directory = os.path.abspath(os.path.join(current_directory, os.pardir))
os.chdir(parent_directory)

from src.paths import *
import pandas as pd
import numpy as np

# Set PD display options to show all columns and column width
pd.set_option('display.max_columns', None)
pd.set_option('display.max_colwidth', 1000)
pd.set_option('display.max_rows', 500)
pd.set_option('display.width', 1000)

#heavy_manufacturing = pd.read_csv(LAND_USE_DIR / 'Complete_Heavy_Manufacturing_Codes.csv', index_col = [0])

SL = pd.read_csv(NORMALIZED_DATA_DIR / 'sales_list_normalized.csv', index_col = [0])
LR = pd.read_csv(NORMALIZED_DATA_DIR / 'local_roll_normalized.csv', index_col = [0])
DS = pd.read_csv(NORMALIZED_DATA_DIR / 'secured_basic_normalized.csv', index_col = [0])
HZ = pd.read_csv(NORMALIZED_DATA_DIR / 'hazards_normalized.csv', index_col = [0])
LU = pd.read_csv(NORMALIZED_DATA_DIR / 'land_use_normalized.csv', index_col = [0])

# Save the data to the NORMALIZED_DATA_DIR
from src.paths import FRONTEND_DIR

SL.to_csv(FRONTEND_DIR / 'sales_list_frontend.csv')
LR.to_csv(FRONTEND_DIR / 'local_roll_frontend.csv')
DS.to_csv(FRONTEND_DIR / 'secured_basic_frontend.csv')
HZ.to_csv(FRONTEND_DIR / 'hazards_frontend.csv')
LU.to_csv(FRONTEND_DIR / 'land_use_frontend.csv')
